requirements.txt
bs4
python-dateutil
requests
